refactor: Rename 'encoder' to 'encoders'
